+++ arm9/source/themes/custom/CustomTopCoverInfo.h
@@ -1,0 +1,14 @@
+#pragma once
+#include "core/math/Point.h"
+
+class CustomTopCoverInfo
+{
+public:
+    CustomTopCoverInfo(const Point& position)
+        : _position(position) { }
+
+    const Point& GetPosition() const { return _position; }
+
+private:
+    Point _position;
+};